requirements.txt
tensorflow==1.13.1
pillow
opencv-python
scikit-image
opendr
cython==0.29.7
numpy==1.16.1
scipy==1.2.1
openpyxl==2.6.2
ecos==2.0.7.post1
vtk==8.1.2
traits==5.1.1
cvxpy
fancyimpute==0.3.2
mayavi==4.7.1
pyopengl==3.1.0
pyqt5==5.10.1
scikit-learn==0.21.3
